Offer model also has "amounts"
The offer model also has amounts. Since we have a functionfor this
already, I refactored it into a `deserialize` module along with the
from_str function.
