+++ resources/src/lib.rs
@@ -7,7 +7,7 @@ extern crate serde;
 extern crate serde_derive;
 extern crate serde_json;
 
-mod utils;
+mod deserialize;
 mod account;
 mod asset;
